[INLONG-8360][Manager] Rename queue resource operator related-classes name
