+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/resource/queue/pulsar/PulsarQueueResourceOperator.java
@@ -57,7 +57,7 @@ import java.util.List;
  */
 @Slf4j
 @Service
-public class PulsarResourceOperator implements QueueResourceOperator {
+public class PulsarQueueResourceOperator implements QueueResourceOperator {
 
     /**
      * The name rule for Pulsar subscription: clusterTag_topicName_sinkId_consumer_group
